Started doing controller/service revision

+++ STAREvents.Services/EventHelperService.cs
@@ -3,9 +3,6 @@ using STAREvents.Data.Models;
 using STAREvents.Data.Repository.Interfaces;
 using STAREvents.Services.Data.Interfaces;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace STAREvents.Services.Data
@@ -15,20 +12,31 @@ namespace STAREvents.Services.Data
         private readonly IRepository<Event, object> eventRepository;
         private readonly IRepository<UserEventAttendance, object> attendanceRepository;
         private readonly UserManager<ApplicationUser> userManager;
-        public EventHelperService(IRepository<Event, object> _eventRepository,
-            IRepository<UserEventAttendance, object> _attendanceRepository,
-            UserManager<ApplicationUser> _userManager)
-        {
-            this.userManager = _userManager;
-            this.eventRepository = _eventRepository;
-            this.attendanceRepository = _attendanceRepository;
-        }
-        public async Task JoinEventAsync(Guid eventId, string userName)
-        {
-            var user = await userManager.FindByNameAsync(userName);
-            var eventEntity = await eventRepository.GetByIdAsync(eventId);
-            var attendance = await attendanceRepository.FirstOrDefaultAsync(a => a.EventId == eventId && a.UserId == user.Id);
 
+        public EventHelperService(IRepository<Event, object> eventRepository,
+            IRepository<UserEventAttendance, object> attendanceRepository,
+            UserManager<ApplicationUser> userManager)
+        {
+            this.userManager = userManager;
+            this.eventRepository = eventRepository;
+            this.attendanceRepository = attendanceRepository;
+        }
+
+        public async Task JoinEventAsync(Guid eventId, Guid userId)
+        {
+            var user = await userManager.FindByIdAsync(userId.ToString());
+            if (user == null)
+            {
+                throw new KeyNotFoundException("User not found");
+            }
+
+            var eventEntity = await eventRepository.GetByIdAsync(eventId);
+            if (eventEntity == null)
+            {
+                throw new KeyNotFoundException("Event not found");
+            }
+
+            var attendance = await attendanceRepository.FirstOrDefaultAsync(a => a.EventId == eventId && a.UserId == user.Id);
             if (attendance == null)
             {
                 var newAttendance = new UserEventAttendance
@@ -43,12 +51,21 @@ namespace STAREvents.Services.Data
             }
         }
 
-        public async Task LeaveEventAsync(Guid eventId, string userName)
+        public async Task LeaveEventAsync(Guid eventId, Guid userId)
         {
-            var user = await userManager.FindByNameAsync(userName);
-            var eventEntity = await eventRepository.GetByIdAsync(eventId);
-            var attendance = await attendanceRepository.FirstOrDefaultAsync(a => a.EventId == eventId && a.UserId == user.Id);
+            var user = await userManager.FindByIdAsync(userId.ToString());
+            if (user == null)
+            {
+                throw new KeyNotFoundException("User not found");
+            }
 
+            var eventEntity = await eventRepository.GetByIdAsync(eventId);
+            if (eventEntity == null)
+            {
+                throw new KeyNotFoundException("Event not found");
+            }
+
+            var attendance = await attendanceRepository.FirstOrDefaultAsync(a => a.EventId == eventId && a.UserId == user.Id);
             if (attendance != null)
             {
                 await attendanceRepository.DeleteAsync(attendance);
@@ -58,3 +75,8 @@ namespace STAREvents.Services.Data
         }
     }
 }
+
+
+
+
+
